npc_human_medic healthkit drop implementation for mp
It is disabled in orig code if more then 1 players can be on server.
The only condition that checks the orig code is player's heath, if health < 50 - the medic wil always drop healthkit on death, even if the killer isn't the player. This is why I decided to check health for every alive player.
Adds ENTPATCH_BM_MEDIC macros.
Tested with dummy bots (before I added fake client check), seems to work properly.